- updated and modernized unistd.h, fcntl.h sys/stat.h   splitted off system dependent declarations into bits/*.h   files - lot of type corrections all over the place due to the header   cleanup - added lot of missing weak symbols - rework of the Fstat64/Fxattr system call access; there is now   a global system call wrapper that optimally wrap stat/fstat   to the running system; replaced any direct Fxattr call with this   wrapper

+++ mintlib/getcwd.c
@@ -24,7 +24,7 @@ extern char _rootdir;	/* in main.c: user's preferred root directory */
 #endif
 
 char *
-__getcwd(char *buf, int size)
+__getcwd (char *buf, size_t size)
 {
 	const int len = (size > 0 ? size : PATH_MAX) + 16;
 	char _path[len]; /* XXX non ANSI */
