+++ unix/fork.c
@@ -1,23 +1,21 @@
 /* public domain fork() for MiNT */
 
-#include <unistd.h>
 #include <errno.h>
-#include <mintbind.h>
+#include <unistd.h>
+#include <mint/mintbind.h>
 
 #include "lib.h"
 
 int
 __fork (void)
 {
-	int r;
+	long r;
 
-	r = (int)Pfork();
+	r = Pfork ();
 	if (r < 0) {
 		__set_errno (-r);
 		return -1;
 	}
 	return r;
 }
-
 weak_alias (__fork, fork)
-
